Corrección de bug en cuadro de inscripciones

+++ php/becas/inscripcion/ci_inscripcion.php
@@ -7,11 +7,14 @@ class ci_inscripcion extends becas_ci
 		//si no existen convocatorias con inscripcion abierta, elimino el evento 'agregar (Nueva Inscripcion'
 		if( ! toba::consulta_php('co_convocatoria_beca')->existen_convocatorias_vigentes()){
 			$this->pantalla()->eliminar_evento('agregar');
-			$this->dep('cuadro')->agregar_notificacion('No existen convocatorias con periodo de inscripción abierto');
+			$this->dep('cuadro')->agregar_notificacion('No existen convocatorias con periodo de inscripci? abierto');
 		}
 		//si el usuario es becario, solo puede ver sus propias inscripciones
 		if(!in_array('admin',toba::usuario()->get_perfiles_funcionales())){
-			$this->pantalla('pant_seleccion')->eliminar_dep('form_filtro');
+			if($this->existe_dependencia('form_filtro')){
+				$this->pantalla('pant_seleccion')->eliminar_dep('form_filtro');	
+			}
+			
 		}
 	}
 	//---- Cuadro Inscripciones ------------------------------------------------------------
@@ -30,11 +33,11 @@ class ci_inscripcion extends becas_ci
 
 	function evt__cuadro__seleccion($datos)
 	{
-		//se carga la relación  de "Alumno"
+		//se carga la relaci?  de "Alumno"
 		$alumno = array('nro_documento'=>$datos['nro_documento']);
 		$this->get_datos('alumno')->cargar($alumno);
 
-		//se cargan los detalles de la inscripción
+		//se cargan los detalles de la inscripci?
 		$this->get_datos('inscripcion')->cargar($datos);
 
 		$this->set_pantalla('pant_edicion');
@@ -134,7 +137,7 @@ class ci_inscripcion extends becas_ci
 					break;
 				
 				default:
-					toba::notificacion()->agregar('Ocurrió un error inesperado al intentar guardar la inscripción. por favor, comuniquese con la Secretaría General de Ciencia y Técnica para solucionarlo (cyt.unne@gmail.com). Código de error: '.$e->get_mensaje_motor());	
+					toba::notificacion()->agregar('Ocurri?un error inesperado al intentar guardar la inscripci?. por favor, comuniquese con la Secretar? General de Ciencia y T?nica para solucionarlo (cyt.unne@gmail.com). C?igo de error: '.$e->get_mensaje_motor());	
 					break;
 			}
 		}catch(toba_error $e){
@@ -187,7 +190,7 @@ class ci_inscripcion extends becas_ci
 			}
 		}
 		if( ($inscripcion['nro_documento_codir'] || $inscripcion['nro_documento_codir']) &&  !$inscripcion['justif_codirector']){
-			$faltantes['Co-Director'][] = 'Justificación del Co-Director/Sub-Director';
+			$faltantes['Co-Director'][] = 'Justificaci? del Co-Director/Sub-Director';
 		}
 		if(!$plan_trabajo['doc_probatoria']){
 			$faltantes['Plan Trabajo'][] = 'Archivo PDF con el plan de trabajo';
@@ -205,12 +208,12 @@ class ci_inscripcion extends becas_ci
 		if(count($faltantes) > 0){
 			$mensaje = "<ul>";
 			foreach ($faltantes as $pantalla => $campos) {
-				$mensaje .= "<li>Sección '$pantalla': ".implode(', ',$campos)."</li>";
+				$mensaje .= "<li>Secci? '$pantalla': ".implode(', ',$campos)."</li>";
 			}
 			$mensaje .= "</ul>";
 			throw new toba_error('Faltan datos por completar: '.$mensaje);
 		}else{
-			toba::notificacion()->agregar('Se ha cerrado correctamente la solicitud. En la parte inferior de esta pantalla puede descargar el comprobante de inscripción, el cual debe ser entregado a la SGCyT.','info');
+			toba::notificacion()->agregar('Se ha cerrado correctamente la solicitud. En la parte inferior de esta pantalla puede descargar el comprobante de inscripci?, el cual debe ser entregado a la SGCyT.','info');
 			$this->get_datos('inscripcion','inscripcion_conv_beca')->set(array('estado'=>'C'));
 			$this->evt__guardar();
 		}
